@@ -1,70 +1,56 @@
-var observable_1 = require("data/observable");
-var tnsfx = require('nativescript-effects');
-var MainPageController = (function (_super) {
-    __extends(MainPageController, _super);
-    function MainPageController() {
-        _super.call(this);
-        this.counter = 42;
-        this.set("message", "number of taps");
-    }
-    MainPageController.prototype.pageLoaded = function (args) {
-        var _this = this;
-        this.page = args.object;
-        this.page.bindingContext = this;
-        this.lblTitle = this.page.getViewById('lblTitle');
-        this.btnTap = this.page.getViewById('btnTap');
-        this.lblCounter = this.page.getViewById('lblCounter');
-        this.lblTitle.slideDown(1000, -50)
+/*
+In NativeScript, a file with the same name as an XML file is known as
+a code-behind file. The code-behind is a great place to place your view
+logic, and to set up your page’s data binding.
+*/
+"use strict";
+var main_view_model_1 = require("./main-view-model");
+var lblTitle;
+var btnTap;
+var lblCounter;
+// Event handler for Page "navigatingTo" event attached in main-page.xml
+function navigatingTo(args) {
+    /*
+    This gets a reference this page’s <Page> UI component. You can
+    view the API reference of the Page to see what’s available at
+    https://docs.nativescript.org/api-reference/classes/_ui_page_.page.html
+    */
+    var page = args.object;
+    /*
+    A page’s bindingContext is an object that should be used to perform
+    data binding between XML markup and TypeScript code. Properties
+    on the bindingContext can be accessed using the {{ }} syntax in XML.
+    In this example, the {{ message }} and {{ onTap }} bindings are resolved
+    against the object returned by createViewModel().
+
+    You can learn more about data binding in NativeScript at
+    https://docs.nativescript.org/core-concepts/data-binding.
+    */
+    page.bindingContext = new main_view_model_1.HelloWorldModel();
+    lblTitle = page.getViewById('lblTitle');
+    btnTap = page.getViewById('btnTap');
+    lblCounter = page.getViewById('lblCounter');
+    lblTitle.slideDown(1000, -50)
+        .then(function () {
+        btnTap.fadeIn(5000)
             .then(function () {
-            _this.btnTap.fadeIn(5000)
-                .then(function () {
-                _this.lblTitle.slideUp(200, -50);
-                _this.btnTap.spring(2000, {
-                    translate: {
-                        x: 0,
-                        y: -50
-                    },
-                    scale: {
-                        x: 3.0,
-                        y: 0.5
-                    },
-                    delay: 0,
-                    dampingRatio: 0.3,
-                    velocity: 2.0,
-                    options: null
-                });
+            lblTitle.slideUp(200, -50);
+            btnTap.spring(2000, {
+                translate: {
+                    x: 0,
+                    y: -50
+                },
+                scale: {
+                    x: 3.0,
+                    y: 0.5
+                },
+                delay: 0,
+                dampingRatio: 0.3,
+                velocity: 2.0,
+                options: null
             });
         });
-    };
-    MainPageController.prototype.randomLocNum = function () {
-        return Math.floor(Math.random() * 100) + 1;
-    };
-    MainPageController.prototype.randomScaleNum = function () {
-        return Math.floor(Math.random() * 30) / 10 + 1;
-    };
-    MainPageController.prototype.tapAction = function () {
-        this.counter++;
-        this.set("message", this.counter + " taps");
-        this.lblCounter.hide();
-        /*
-        this.btnTap.spring(2000, {
-            scale:{ x: this.randomScaleNum(), y: this.randomScaleNum() },
-                    delay: 0,
-                    dampingRatio: 0.6,
-                    velocity: 2.5,
-                    options: null
-        });
-        
-        this.lblCounter.spring(5000, {
-                    translate:{ x: this.randomLocNum(), y: this.randomLocNum() },
-                    delay: 0,
-                    dampingRatio: 0.3,
-                    velocity: 6.0,
-                    options: null});
-                    */
-    };
-    return MainPageController;
-})(observable_1.Observable);
-var mpc = new MainPageController();
-exports.pageLoaded = function (args) { return mpc.pageLoaded(args); };
-exports.tapAction = function () { return mpc.tapAction(); };
+    });
+}
+exports.navigatingTo = navigatingTo;
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoibWFpbi1wYWdlLmpzIiwic291cmNlUm9vdCI6IiIsInNvdXJjZXMiOlsibWFpbi1wYWdlLnRzIl0sIm5hbWVzIjpbXSwibWFwcGluZ3MiOiJBQUFBOzs7O0VBSUU7O0FBTUYscURBQW9EO0FBR3BELElBQUksUUFBZSxDQUFDO0FBQ3BCLElBQUksTUFBYyxDQUFDO0FBQ25CLElBQUksVUFBaUIsQ0FBQztBQUV0Qix3RUFBd0U7QUFDeEUsc0JBQTZCLElBQWU7SUFDeEM7Ozs7TUFJRTtJQUNGLElBQUksSUFBSSxHQUFTLElBQUksQ0FBQyxNQUFNLENBQUM7SUFFN0I7Ozs7Ozs7OztNQVNFO0lBQ0YsSUFBSSxDQUFDLGNBQWMsR0FBRyxJQUFJLGlDQUFlLEVBQUUsQ0FBQztJQUU1QyxRQUFRLEdBQVUsSUFBSSxDQUFDLFdBQVcsQ0FBQyxVQUFVLENBQUMsQ0FBQztJQUMvQyxNQUFNLEdBQVcsSUFBSSxDQUFDLFdBQVcsQ0FBQyxRQUFRLENBQUMsQ0FBQztJQUM1QyxVQUFVLEdBQVUsSUFBSSxDQUFDLFdBQVcsQ0FBQyxZQUFZLENBQUMsQ0FBQztJQUVuRCxRQUFRLENBQUMsU0FBUyxDQUFDLElBQUksRUFBRSxDQUFDLEVBQUUsQ0FBQztTQUN4QixJQUFJLENBQUM7UUFDRixNQUFNLENBQUMsTUFBTSxDQUFDLElBQUksQ0FBQzthQUNkLElBQUksQ0FBQztZQUNGLFFBQVEsQ0FBQyxPQUFPLENBQUMsR0FBRyxFQUFFLENBQUMsRUFBRSxDQUFDLENBQUM7WUFDM0IsTUFBTSxDQUFDLE1BQU0sQ0FBQyxJQUFJLEVBQUU7Z0JBQ2hCLFNBQVMsRUFBRTtvQkFDUCxDQUFDLEVBQUUsQ0FBQztvQkFDSixDQUFDLEVBQUUsQ0FBQyxFQUFFO2lCQUNUO2dCQUNELEtBQUssRUFBRTtvQkFDSCxDQUFDLEVBQUUsR0FBRztvQkFDTixDQUFDLEVBQUUsR0FBRztpQkFDVDtnQkFDRCxLQUFLLEVBQUUsQ0FBQztnQkFDUixZQUFZLEVBQUUsR0FBRztnQkFDakIsUUFBUSxFQUFFLEdBQUc7Z0JBQ2IsT0FBTyxFQUFFLElBQUk7YUFDaEIsQ0FBQyxDQUFDO1FBQ1AsQ0FBQyxDQUFDLENBQUM7SUFDWCxDQUFDLENBQ0EsQ0FBQztBQUVWLENBQUM7QUEvQ0Qsb0NBK0NDIiwic291cmNlc0NvbnRlbnQiOlsiLypcbkluIE5hdGl2ZVNjcmlwdCwgYSBmaWxlIHdpdGggdGhlIHNhbWUgbmFtZSBhcyBhbiBYTUwgZmlsZSBpcyBrbm93biBhc1xuYSBjb2RlLWJlaGluZCBmaWxlLiBUaGUgY29kZS1iZWhpbmQgaXMgYSBncmVhdCBwbGFjZSB0byBwbGFjZSB5b3VyIHZpZXdcbmxvZ2ljLCBhbmQgdG8gc2V0IHVwIHlvdXIgcGFnZeKAmXMgZGF0YSBiaW5kaW5nLlxuKi9cblxuaW1wb3J0IHsgRXZlbnREYXRhIH0gZnJvbSAnZGF0YS9vYnNlcnZhYmxlJztcbmltcG9ydCB7IFBhZ2UgfSBmcm9tICd1aS9wYWdlJztcbmltcG9ydCB7IEJ1dHRvbiB9IGZyb20gJ3VpL2J1dHRvbic7XG5pbXBvcnQgeyBMYWJlbCB9IGZyb20gJ3VpL2xhYmVsJztcbmltcG9ydCB7IEhlbGxvV29ybGRNb2RlbCB9IGZyb20gJy4vbWFpbi12aWV3LW1vZGVsJztcblxuXG52YXIgbGJsVGl0bGU6IExhYmVsO1xudmFyIGJ0blRhcDogQnV0dG9uO1xudmFyIGxibENvdW50ZXI6IExhYmVsO1xuXG4vLyBFdmVudCBoYW5kbGVyIGZvciBQYWdlIFwibmF2aWdhdGluZ1RvXCIgZXZlbnQgYXR0YWNoZWQgaW4gbWFpbi1wYWdlLnhtbFxuZXhwb3J0IGZ1bmN0aW9uIG5hdmlnYXRpbmdUbyhhcmdzOiBFdmVudERhdGEpIHtcbiAgICAvKlxuICAgIFRoaXMgZ2V0cyBhIHJlZmVyZW5jZSB0aGlzIHBhZ2XigJlzIDxQYWdlPiBVSSBjb21wb25lbnQuIFlvdSBjYW5cbiAgICB2aWV3IHRoZSBBUEkgcmVmZXJlbmNlIG9mIHRoZSBQYWdlIHRvIHNlZSB3aGF04oCZcyBhdmFpbGFibGUgYXRcbiAgICBodHRwczovL2RvY3MubmF0aXZlc2NyaXB0Lm9yZy9hcGktcmVmZXJlbmNlL2NsYXNzZXMvX3VpX3BhZ2VfLnBhZ2UuaHRtbFxuICAgICovXG4gICAgbGV0IHBhZ2UgPSA8UGFnZT5hcmdzLm9iamVjdDtcblxuICAgIC8qXG4gICAgQSBwYWdl4oCZcyBiaW5kaW5nQ29udGV4dCBpcyBhbiBvYmplY3QgdGhhdCBzaG91bGQgYmUgdXNlZCB0byBwZXJmb3JtXG4gICAgZGF0YSBiaW5kaW5nIGJldHdlZW4gWE1MIG1hcmt1cCBhbmQgVHlwZVNjcmlwdCBjb2RlLiBQcm9wZXJ0aWVzXG4gICAgb24gdGhlIGJpbmRpbmdDb250ZXh0IGNhbiBiZSBhY2Nlc3NlZCB1c2luZyB0aGUge3sgfX0gc3ludGF4IGluIFhNTC5cbiAgICBJbiB0aGlzIGV4YW1wbGUsIHRoZSB7eyBtZXNzYWdlIH19IGFuZCB7eyBvblRhcCB9fSBiaW5kaW5ncyBhcmUgcmVzb2x2ZWRcbiAgICBhZ2FpbnN0IHRoZSBvYmplY3QgcmV0dXJuZWQgYnkgY3JlYXRlVmlld01vZGVsKCkuXG5cbiAgICBZb3UgY2FuIGxlYXJuIG1vcmUgYWJvdXQgZGF0YSBiaW5kaW5nIGluIE5hdGl2ZVNjcmlwdCBhdFxuICAgIGh0dHBzOi8vZG9jcy5uYXRpdmVzY3JpcHQub3JnL2NvcmUtY29uY2VwdHMvZGF0YS1iaW5kaW5nLlxuICAgICovXG4gICAgcGFnZS5iaW5kaW5nQ29udGV4dCA9IG5ldyBIZWxsb1dvcmxkTW9kZWwoKTtcblxuICAgIGxibFRpdGxlID0gPExhYmVsPnBhZ2UuZ2V0Vmlld0J5SWQoJ2xibFRpdGxlJyk7XG4gICAgYnRuVGFwID0gPEJ1dHRvbj5wYWdlLmdldFZpZXdCeUlkKCdidG5UYXAnKTtcbiAgICBsYmxDb3VudGVyID0gPExhYmVsPnBhZ2UuZ2V0Vmlld0J5SWQoJ2xibENvdW50ZXInKTtcblxuICAgIGxibFRpdGxlLnNsaWRlRG93bigxMDAwLCAtNTApXG4gICAgICAgIC50aGVuKCgpID0+IHtcbiAgICAgICAgICAgIGJ0blRhcC5mYWRlSW4oNTAwMClcbiAgICAgICAgICAgICAgICAudGhlbigoKSA9PiB7XG4gICAgICAgICAgICAgICAgICAgIGxibFRpdGxlLnNsaWRlVXAoMjAwLCAtNTApO1xuICAgICAgICAgICAgICAgICAgICBidG5UYXAuc3ByaW5nKDIwMDAsIHtcbiAgICAgICAgICAgICAgICAgICAgICAgIHRyYW5zbGF0ZToge1xuICAgICAgICAgICAgICAgICAgICAgICAgICAgIHg6IDAsXG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgeTogLTUwXG4gICAgICAgICAgICAgICAgICAgICAgICB9LFxuICAgICAgICAgICAgICAgICAgICAgICAgc2NhbGU6IHtcbiAgICAgICAgICAgICAgICAgICAgICAgICAgICB4OiAzLjAsXG4gICAgICAgICAgICAgICAgICAgICAgICAgICAgeTogMC41XG4gICAgICAgICAgICAgICAgICAgICAgICB9LFxuICAgICAgICAgICAgICAgICAgICAgICAgZGVsYXk6IDAsXG4gICAgICAgICAgICAgICAgICAgICAgICBkYW1waW5nUmF0aW86IDAuMyxcbiAgICAgICAgICAgICAgICAgICAgICAgIHZlbG9jaXR5OiAyLjAsXG4gICAgICAgICAgICAgICAgICAgICAgICBvcHRpb25zOiBudWxsXG4gICAgICAgICAgICAgICAgICAgIH0pO1xuICAgICAgICAgICAgICAgIH0pO1xuICAgICAgICB9XG4gICAgICAgICk7XG5cbn0iXX0=
